backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
requests
python-dotenv